refactor(state): rename new_state to state
