+++ backend/crates/desktop/src/state.rs
